Adds Deploy Error Notification
The core change here is that we refactor the Notifier.deploySuccess to
return post message arguments, instead of a TaskEither, and actually
post the message back up at the top layer in `main`.

This allows us to create Notifier.deployError following the same pattern
and have all of the types line up cleanly.